Sign flashable ZIPs
For some reason, some recoveries include "unzip" binaries that can't handle
signed ZIP archives. This is fixed with this BusyBox commit:
https://github.com/CyanogenMod/android_external_busybox/commit/2b3b6440c6b30451489f77bd4edcaff85011ed3c

As we use our special BusyBox executable (including this fix) to unzip the
files, we don't have to rely on the recovery anymore. So the archives can be
signed now.