Add the possiblity to flash files automatically
This requires a working "adb" connection to the device. The ZIP file is pushed
to the devices, the BusyBox executable as well. Then the installation procedure
is performed. This is meant for test installations on a running system.

NOTE: To enable network mode, execute "setprop persist.adb.tcp.port 5555".
After a reboot, you can use "adb connect <IP>" and then use this script.