Rename .odex.gz files to .odex.gz.xposed
On some ROMs, the odex files are compressed with gzip. The installd daemin
on these ROMs directs all dexopt requests to patchoat, but we need it to go
through dex2oat. Instead of patching installd, we simply hide the compressed
files from it and handle them properly inside dex2oat.