Support for host builds
If "host" or "hostd" (for debug builds) is given as platform, dex2oat
and oatdump including their dependencies will be built for the host,
which is assumed to be linux-x86. This can be helpful for debugging
low-level issues without having to flash new files to a device.

These builds are not included when using "all" for the platform.