Require that the installer is aware of FBE on Nougat
File-Based Encryption forces us to use /data/user_de for storing module
lists and other settings. This path is hard-coded in several places,
including the Xposed Installer. However, if the user didn't update the
installer, they'll still be using the legacy path. So add a requirement
to the xposed.prop file that will be checked by the installer and will
ask the user to update.